Utilize stages to seperate the build of the target libraries and the toolchain (#14)
The primary goal with this merge request is to split the build of libraries from every OS to only being done in an Ubuntu container. This split allows for the actual libraries to be 1:1 on all hosts, so there is no question of if autoconf made any implicit choices to work around the host. Also, the hash of the libraries should be the same for all hosts.

Another change was the normalization of the sysroot. This makes every target sysroot have a similar layout to N.I. Linux which helps with the maintainability of the build scripts.