[GCC 7-10; Darwin] Workaround genconditions bug with M1.
Although the conditions that are referenced within the patch files
should not be doing anything with
Apple Clang 13, it is able to fix the compile failure when the M1 is
the host.

This patch was originally found at
https://github.com/crosstool-ng/crosstool-ng/pull/1638 which pulls from
https://gcc.gnu.org/bugzilla/show_bug.cgi?id=92061. This patch was meant
to work around a bug in Clang 9 or lower, but fixes a Clang 13 issue.

It gets weirder because the core issue is not present with GCC 11 even
though the contents of the patch are not applied.